Artist-mini-app/bot/requirements.txt
aiogram==3.19.0
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3